+++ src/main/java/com/coveo/spillway/Spillway.java
@@ -25,6 +25,15 @@ package com.coveo.spillway;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.coveo.spillway.exception.SpillwayLimitExceededException;
+import com.coveo.spillway.limit.Limit;
+import com.coveo.spillway.limit.LimitBuilder;
+import com.coveo.spillway.limit.LimitDefinition;
+import com.coveo.spillway.limit.LimitKey;
+import com.coveo.spillway.limit.LimitUsageStorage;
+import com.coveo.spillway.storage.AddAndGetRequest;
+import com.coveo.spillway.trigger.LimitTrigger;
+
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Arrays;
