+++ src/main/java/com/coveo/spillway/limit/LimitUsageStorage.java
@@ -20,9 +20,11 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.coveo.spillway;
+package com.coveo.spillway.limit;
 
 import org.apache.commons.lang3.tuple.Pair;
+
+import com.coveo.spillway.storage.AddAndGetRequest;
 
 import java.time.Duration;
 import java.time.Instant;
