+++ src/main/java/com/coveo/spillway/storage/AddAndGetRequest.java
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.coveo.spillway;
+package com.coveo.spillway.storage;
 
 import java.time.Duration;
 import java.time.Instant;
