+++ src/test/java/com/coveo/spillway/storage/AsyncLimitUsageStorageTest.java
@@ -20,8 +20,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.coveo.spillway;
+package com.coveo.spillway.storage;
 
+import com.coveo.spillway.limit.LimitKey;
+import com.coveo.spillway.limit.LimitUsageStorage;
+import com.coveo.spillway.storage.AsyncLimitUsageStorage;
+import com.coveo.spillway.storage.AddAndGetRequest.Builder;
 import com.google.common.collect.ImmutableMap;
 
 import org.junit.Before;
